fix(multipart-parser): prevent infinite loop on an over-long boundary
The Boyer-Moore-Horspool skip table in buffer-search.ts is a Uint8Array
filled with needle.length. A multipart boundary of 252 or more characters
makes the search needle 256 bytes, so the fill wraps to 0, the default skip
becomes 0, and the search loop never advances — an infinite-loop denial of
service on runtimes that use the JS search path (Bun, Deno, browsers; Node
uses the native Buffer.indexOf and was unaffected).

Switch the skip table to Uint32Array, which cannot wrap for any pattern
length a JavaScript string can produce (V8/JSC cap string length below 2^29,
well under 2^32). No behavior change for valid input; fixes every entry
point and every createSearch caller.

See GHSA-hpwp-7w8r-373j.

+++ packages/multipart-parser/src/lib/buffer-search.ts
@@ -23,7 +23,11 @@ export function createSearch(pattern: string): SearchFunction {
     search = (haystack, start = 0) => BufferClass.prototype.indexOf.call(haystack, needle, start)
   } else {
     let needleEnd = needle.length - 1
-    let skipTable = new Uint8Array(256).fill(needle.length)
+    // Uint32Array, not Uint8Array: for a needle of 256+ bytes the fill value would wrap a
+    // Uint8Array element to 0, making the default skip 0 so the loop below never advances
+    // (infinite loop). needle.length is bounded by the pattern string length, well under
+    // 2^32, so a Uint32 skip can never wrap.
+    let skipTable = new Uint32Array(256).fill(needle.length)
     for (let i = 0; i < needleEnd; ++i) {
       skipTable[needle[i]] = needleEnd - i
     }
